requirements.txt
playwright
python-dateutil